Initial commit
This code has been developed over several years for the BYU Mechanical Engineering 537 class. This code includes templates for all of the homework that we complete throughout the class.

Still need a requirements.txt file.  This code is derived from a library written in collaboration with John Morrell. But with specific parts excised to make learning exercises.